Get a response from the GitHub API.

Get a response from the GitHub API.

gh_get(path, ref = "master")

Arguments

path
A string formatted as "owner/repo/subdir_1/subdir_2/subdir_n".

ref
Branch name.

Value

A list.

See also

Examples

gh_get("r-lib/usethis")

#> [
#>   {
#>     "name": ".Rbuildignore",
#>     "path": ".Rbuildignore",
#>     "sha": "6fbf251f1a6100e147fb2aa7ef55993e0c62e526",
#>     "size": 289,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/.Rbuildignore?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/.Rbuildignore",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/6fbf251f1a6100e147fb2aa7ef55993e0c62e526",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/.Rbuildignore",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/.Rbuildignore?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/6fbf251f1a6100e147fb2aa7ef55993e0c62e526",
#>       "html": "https://github.com/r-lib/usethis/blob/master/.Rbuildignore"
#>     }
#>   },
#>   {
#>     "name": ".covrignore",
#>     "path": ".covrignore",
#>     "sha": "e18d9803cc6ff749e3456f4428265701136058f7",
#>     "size": 46,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/.covrignore?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/.covrignore",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/e18d9803cc6ff749e3456f4428265701136058f7",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/.covrignore",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/.covrignore?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/e18d9803cc6ff749e3456f4428265701136058f7",
#>       "html": "https://github.com/r-lib/usethis/blob/master/.covrignore"
#>     }
#>   },
#>   {
#>     "name": ".github",
#>     "path": ".github",
#>     "sha": "9e3a7fce4bffd08eb24e9ce6c06d934674d58d26",
#>     "size": 0,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/.github?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/tree/master/.github",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/trees/9e3a7fce4bffd08eb24e9ce6c06d934674d58d26",
#>     "download_url": {},
#>     "type": "dir",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/.github?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/trees/9e3a7fce4bffd08eb24e9ce6c06d934674d58d26",
#>       "html": "https://github.com/r-lib/usethis/tree/master/.github"
#>     }
#>   },
#>   {
#>     "name": ".gitignore",
#>     "path": ".gitignore",
#>     "sha": "10c71b274d20eba2b207fa4b547216da84ebaf23",
#>     "size": 43,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/.gitignore?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/.gitignore",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/10c71b274d20eba2b207fa4b547216da84ebaf23",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/.gitignore",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/.gitignore?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/10c71b274d20eba2b207fa4b547216da84ebaf23",
#>       "html": "https://github.com/r-lib/usethis/blob/master/.gitignore"
#>     }
#>   },
#>   {
#>     "name": ".travis.yml",
#>     "path": ".travis.yml",
#>     "sha": "1178b26f5cd8bde9d86efe933c16d18f0cbab9b0",
#>     "size": 501,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/.travis.yml?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/.travis.yml",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/1178b26f5cd8bde9d86efe933c16d18f0cbab9b0",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/.travis.yml",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/.travis.yml?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/1178b26f5cd8bde9d86efe933c16d18f0cbab9b0",
#>       "html": "https://github.com/r-lib/usethis/blob/master/.travis.yml"
#>     }
#>   },
#>   {
#>     "name": "DESCRIPTION",
#>     "path": "DESCRIPTION",
#>     "sha": "1c9e6bc677324b1f629b06d96289060bdfdfd5f8",
#>     "size": 1322,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/DESCRIPTION?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/DESCRIPTION",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/1c9e6bc677324b1f629b06d96289060bdfdfd5f8",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/DESCRIPTION",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/DESCRIPTION?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/1c9e6bc677324b1f629b06d96289060bdfdfd5f8",
#>       "html": "https://github.com/r-lib/usethis/blob/master/DESCRIPTION"
#>     }
#>   },
#>   {
#>     "name": "LICENSE.md",
#>     "path": "LICENSE.md",
#>     "sha": "0a7453cd57c67c97804da7e54e353b1b8719e2a0",
#>     "size": 34877,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/LICENSE.md?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/LICENSE.md",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/0a7453cd57c67c97804da7e54e353b1b8719e2a0",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/LICENSE.md",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/LICENSE.md?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/0a7453cd57c67c97804da7e54e353b1b8719e2a0",
#>       "html": "https://github.com/r-lib/usethis/blob/master/LICENSE.md"
#>     }
#>   },
#>   {
#>     "name": "NAMESPACE",
#>     "path": "NAMESPACE",
#>     "sha": "35f736bb6e283696972e004a8e3de614ee6cfad1",
#>     "size": 3139,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/NAMESPACE?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/NAMESPACE",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/35f736bb6e283696972e004a8e3de614ee6cfad1",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/NAMESPACE",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/NAMESPACE?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/35f736bb6e283696972e004a8e3de614ee6cfad1",
#>       "html": "https://github.com/r-lib/usethis/blob/master/NAMESPACE"
#>     }
#>   },
#>   {
#>     "name": "NEWS.md",
#>     "path": "NEWS.md",
#>     "sha": "f01b7b4b3c5e0c8794f2027ac89464a08ec6f5c1",
#>     "size": 23537,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/NEWS.md?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/NEWS.md",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/f01b7b4b3c5e0c8794f2027ac89464a08ec6f5c1",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/NEWS.md",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/NEWS.md?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/f01b7b4b3c5e0c8794f2027ac89464a08ec6f5c1",
#>       "html": "https://github.com/r-lib/usethis/blob/master/NEWS.md"
#>     }
#>   },
#>   {
#>     "name": "R",
#>     "path": "R",
#>     "sha": "41fc0ec67ca3995d80593713843bd9e99e503557",
#>     "size": 0,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/R?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/tree/master/R",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/trees/41fc0ec67ca3995d80593713843bd9e99e503557",
#>     "download_url": {},
#>     "type": "dir",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/R?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/trees/41fc0ec67ca3995d80593713843bd9e99e503557",
#>       "html": "https://github.com/r-lib/usethis/tree/master/R"
#>     }
#>   },
#>   {
#>     "name": "README.Rmd",
#>     "path": "README.Rmd",
#>     "sha": "57e8bf7ab9e0e1c43a93363defa6838f5b9e09a8",
#>     "size": 3327,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/README.Rmd?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/README.Rmd",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/57e8bf7ab9e0e1c43a93363defa6838f5b9e09a8",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/README.Rmd",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/README.Rmd?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/57e8bf7ab9e0e1c43a93363defa6838f5b9e09a8",
#>       "html": "https://github.com/r-lib/usethis/blob/master/README.Rmd"
#>     }
#>   },
#>   {
#>     "name": "README.md",
#>     "path": "README.md",
#>     "sha": "96afde7383463d11eaba7bfdbe36b4fb9c927e3b",
#>     "size": 5644,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/README.md?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/README.md",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/96afde7383463d11eaba7bfdbe36b4fb9c927e3b",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/README.md",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/README.md?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/96afde7383463d11eaba7bfdbe36b4fb9c927e3b",
#>       "html": "https://github.com/r-lib/usethis/blob/master/README.md"
#>     }
#>   },
#>   {
#>     "name": "_pkgdown.yml",
#>     "path": "_pkgdown.yml",
#>     "sha": "96530a08411cda898bf43ec1b42519af591067ea",
#>     "size": 3004,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/_pkgdown.yml?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/_pkgdown.yml",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/96530a08411cda898bf43ec1b42519af591067ea",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/_pkgdown.yml",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/_pkgdown.yml?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/96530a08411cda898bf43ec1b42519af591067ea",
#>       "html": "https://github.com/r-lib/usethis/blob/master/_pkgdown.yml"
#>     }
#>   },
#>   {
#>     "name": "appveyor.yml",
#>     "path": "appveyor.yml",
#>     "sha": "51ca9664ebe626b7def730abd721dd3f05f10d79",
#>     "size": 994,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/appveyor.yml?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/appveyor.yml",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/51ca9664ebe626b7def730abd721dd3f05f10d79",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/appveyor.yml",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/appveyor.yml?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/51ca9664ebe626b7def730abd721dd3f05f10d79",
#>       "html": "https://github.com/r-lib/usethis/blob/master/appveyor.yml"
#>     }
#>   },
#>   {
#>     "name": "codecov.yml",
#>     "path": "codecov.yml",
#>     "sha": "8f36b6cc1b9cfae27678ce41c2f385032cd40baf",
#>     "size": 176,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/codecov.yml?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/codecov.yml",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/8f36b6cc1b9cfae27678ce41c2f385032cd40baf",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/codecov.yml",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/codecov.yml?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/8f36b6cc1b9cfae27678ce41c2f385032cd40baf",
#>       "html": "https://github.com/r-lib/usethis/blob/master/codecov.yml"
#>     }
#>   },
#>   {
#>     "name": "cran-comments.md",
#>     "path": "cran-comments.md",
#>     "sha": "92fbf2a9b35376785d7a1d220da5025dbf04123f",
#>     "size": 913,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/cran-comments.md?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/cran-comments.md",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/92fbf2a9b35376785d7a1d220da5025dbf04123f",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/cran-comments.md",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/cran-comments.md?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/92fbf2a9b35376785d7a1d220da5025dbf04123f",
#>       "html": "https://github.com/r-lib/usethis/blob/master/cran-comments.md"
#>     }
#>   },
#>   {
#>     "name": "inst",
#>     "path": "inst",
#>     "sha": "be89bc2a25b23958851381a9d9e5e1630a74d751",
#>     "size": 0,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/inst?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/tree/master/inst",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/trees/be89bc2a25b23958851381a9d9e5e1630a74d751",
#>     "download_url": {},
#>     "type": "dir",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/inst?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/trees/be89bc2a25b23958851381a9d9e5e1630a74d751",
#>       "html": "https://github.com/r-lib/usethis/tree/master/inst"
#>     }
#>   },
#>   {
#>     "name": "man",
#>     "path": "man",
#>     "sha": "64170d43bbcc5040379620571ece307e6b49307b",
#>     "size": 0,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/man?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/tree/master/man",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/trees/64170d43bbcc5040379620571ece307e6b49307b",
#>     "download_url": {},
#>     "type": "dir",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/man?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/trees/64170d43bbcc5040379620571ece307e6b49307b",
#>       "html": "https://github.com/r-lib/usethis/tree/master/man"
#>     }
#>   },
#>   {
#>     "name": "pkgdown",
#>     "path": "pkgdown",
#>     "sha": "626c233e9e4010d50cef098ba4eb64c37a1b6151",
#>     "size": 0,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/pkgdown?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/tree/master/pkgdown",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/trees/626c233e9e4010d50cef098ba4eb64c37a1b6151",
#>     "download_url": {},
#>     "type": "dir",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/pkgdown?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/trees/626c233e9e4010d50cef098ba4eb64c37a1b6151",
#>       "html": "https://github.com/r-lib/usethis/tree/master/pkgdown"
#>     }
#>   },
#>   {
#>     "name": "principles.md",
#>     "path": "principles.md",
#>     "sha": "b297b5b4b32047d2672e5edb0cc9bf2a05166a13",
#>     "size": 4932,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/principles.md?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/principles.md",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/b297b5b4b32047d2672e5edb0cc9bf2a05166a13",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/principles.md",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/principles.md?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/b297b5b4b32047d2672e5edb0cc9bf2a05166a13",
#>       "html": "https://github.com/r-lib/usethis/blob/master/principles.md"
#>     }
#>   },
#>   {
#>     "name": "revdep",
#>     "path": "revdep",
#>     "sha": "4a808a9a9dbba7ceebaf872ab61948afc5390b36",
#>     "size": 0,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/revdep?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/tree/master/revdep",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/trees/4a808a9a9dbba7ceebaf872ab61948afc5390b36",
#>     "download_url": {},
#>     "type": "dir",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/revdep?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/trees/4a808a9a9dbba7ceebaf872ab61948afc5390b36",
#>       "html": "https://github.com/r-lib/usethis/tree/master/revdep"
#>     }
#>   },
#>   {
#>     "name": "tests",
#>     "path": "tests",
#>     "sha": "a3ee16063f0c30344b247152ef7fdddc96577d6f",
#>     "size": 0,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/tests?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/tree/master/tests",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/trees/a3ee16063f0c30344b247152ef7fdddc96577d6f",
#>     "download_url": {},
#>     "type": "dir",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/tests?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/trees/a3ee16063f0c30344b247152ef7fdddc96577d6f",
#>       "html": "https://github.com/r-lib/usethis/tree/master/tests"
#>     }
#>   },
#>   {
#>     "name": "usethis.Rproj",
#>     "path": "usethis.Rproj",
#>     "sha": "5d5fb43046cff9f3b89853d937bec0f00b155512",
#>     "size": 384,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/usethis.Rproj?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/blob/master/usethis.Rproj",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/blobs/5d5fb43046cff9f3b89853d937bec0f00b155512",
#>     "download_url": "https://raw.githubusercontent.com/r-lib/usethis/master/usethis.Rproj",
#>     "type": "file",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/usethis.Rproj?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/blobs/5d5fb43046cff9f3b89853d937bec0f00b155512",
#>       "html": "https://github.com/r-lib/usethis/blob/master/usethis.Rproj"
#>     }
#>   },
#>   {
#>     "name": "vignettes",
#>     "path": "vignettes",
#>     "sha": "881e854fa90caa3a57c777b8ea6b64541d24a0e1",
#>     "size": 0,
#>     "url": "https://api.github.com/repos/r-lib/usethis/contents/vignettes?ref=master",
#>     "html_url": "https://github.com/r-lib/usethis/tree/master/vignettes",
#>     "git_url": "https://api.github.com/repos/r-lib/usethis/git/trees/881e854fa90caa3a57c777b8ea6b64541d24a0e1",
#>     "download_url": {},
#>     "type": "dir",
#>     "_links": {
#>       "self": "https://api.github.com/repos/r-lib/usethis/contents/vignettes?ref=master",
#>       "git": "https://api.github.com/repos/r-lib/usethis/git/trees/881e854fa90caa3a57c777b8ea6b64541d24a0e1",
#>       "html": "https://github.com/r-lib/usethis/tree/master/vignettes"
#>     }
#>   }
#> ] 

# The request to GitHub happens only the first time you call gh_get()
system.time(gh_get("r-lib/usethis/R"))

#>    user  system elapsed 
#>   0.036   0.000   0.137 

# Later calls take no time because the first call is memoised
system.time(gh_get("r-lib/usethis/R"))

#>    user  system elapsed 
#>   0.000   0.000   0.001 

ghr_path(gh_get("r-lib/usethis", ref = "gh-pages"))

#>  [1] "CNAME"                "CODE_OF_CONDUCT.html" "CONTRIBUTING.html"   
#>  [4] "ISSUE_TEMPLATE.html"  "LICENSE.html"         "SUPPORT.html"        
#>  [7] "articles"             "authors.html"         "dev"                 
#> [10] "docsearch.css"        "docsearch.js"         "index.html"          
#> [13] "link.svg"             "news"                 "pkgdown.css"         
#> [16] "pkgdown.js"           "pkgdown.yml"          "principles.html"     
#> [19] "reference"            "sitemap.xml"         

# Same
ghr_path(gh_get("r-lib/usethis@gh-pages"))

#>  [1] "CNAME"                "CODE_OF_CONDUCT.html" "CONTRIBUTING.html"   
#>  [4] "ISSUE_TEMPLATE.html"  "LICENSE.html"         "SUPPORT.html"        
#>  [7] "articles"             "authors.html"         "dev"                 
#> [10] "docsearch.css"        "docsearch.js"         "index.html"          
#> [13] "link.svg"             "news"                 "pkgdown.css"         
#> [16] "pkgdown.js"           "pkgdown.yml"          "principles.html"     
#> [19] "reference"            "sitemap.xml"         

ghr_path(gh_get("r-lib/usethis/news@gh-pages"))

#> [1] "news/index.html"